+++ tests/unit/RemoteSourceTimeTravelTest.php
@@ -222,6 +222,7 @@ class RemoteSourceTimeTravelTest extends TestCase {
 		$node = $this->make_remote( 'remote-austin' );
 		Core::$now = 1000.0;
 		$node->fire();
+		$this->drain_connect_queue();
 		$sse = Core::node( 'remote-austin:sse-in' );
 		$this->assertInstanceOf( \CurlHandle::class, $sse->test_get_handle(), 'precondition: connected' );